Create working integration test
