+++ api/app/routers/land_change/tree_cover_loss/tree_cover_loss.py
@@ -1,18 +1,15 @@
 from app.domain.analyzers.tree_cover_loss_analyzer import TreeCoverLossAnalyzer
 from app.domain.compute_engines.compute_engine import (
     ComputeEngine,
+    DuckDbPrecalcQueryService,
     FloxOTFHandler,
     PrecalcHandler,
-    PrecalcQueryService,
 )
 from app.domain.repositories.analysis_repository import AnalysisRepository
 from app.domain.repositories.data_api_aoi_geometry_repository import (
     DataApiAoiGeometryRepository,
 )
 from app.domain.repositories.zarr_dataset_repository import ZarrDatasetRepository
-from app.infrastructure.external_services.data_api_compute_service import (
-    DataApiComputeService,
-)
 from app.infrastructure.persistence.file_system_analysis_repository import (
     FileSystemAnalysisRepository,
 )
@@ -38,26 +35,21 @@ def get_analysis_repository() -> AnalysisRepository:
     return FileSystemAnalysisRepository(ANALYTICS_NAME)
 
 
-def create_compute_engine() -> ComputeEngine:
+def create_analysis_service(request: Request) -> AnalysisService:
     compute_engine = ComputeEngine(
         handler=PrecalcHandler(
-            precalc_query_service=PrecalcQueryService(),
+            precalc_query_service=DuckDbPrecalcQueryService(),
             next_handler=FloxOTFHandler(
                 dataset_repository=ZarrDatasetRepository(),
                 aoi_geometry_repository=DataApiAoiGeometryRepository(),
-                # dask_client=dask_client, TODO how to pass request object?
+                dask_client=request.app.state.dask_client,
             ),
         )
     )
 
-    return compute_engine
-
-
-def create_analysis_service() -> AnalysisService:
-    analysis_repository = FileSystemAnalysisRepository(ANALYTICS_NAME)
     return AnalysisService(
-        analysis_repository=analysis_repository,
-        compute_engine=create_compute_engine(),
+        analysis_repository=get_analysis_repository(),
+        analyzer=TreeCoverLossAnalyzer(compute_engine),
         event=ANALYTICS_NAME,
     )
